Created file for helpers functions
